requirements.txt
streamlit==0.74.1
matplotlib==3.2.2
tqdm==4.47.0
av==8.0.3
opencv-python==2.4.9
Pillow==8.1.0
scikit-video==0.1.1
torch==1.7.1
torchvision==0.8.2
